feat: Enhance ENS integration for StockShield pools
- Implemented dynamic ENS name resolution for pools using wagmi hooks.
- Updated PoolNameDisplay and PoolNameBadge components to display ENS names with verification badges.
- Modified PoolsPage to show ENS names alongside pool details.
- Introduced usePoolENSData hook for fetching ENS metadata including text records.
- Removed caching mechanism in favor of real-time ENS resolution.
- Added auto session notifications in YellowNetworkPanel for improved user experience.
- Refactored usePoolENS and usePoolsENS hooks to streamline ENS data retrieval.

+++ contracts/script/CreatePoolOnly.s.sol
@@ -70,7 +70,7 @@ contract CreatePoolOnly is Script {
         } else {
             // Stock is currency0, USDC is currency1
             // price = 10^-12, sqrt = 10^-6, * 2^96
-            sqrtPriceX96 = uint160((2**96) / 1e6);
+            sqrtPriceX96 = uint160(uint256(2**96) / 1e6);
         }
         console2.log("  sqrtPriceX96:", uint256(sqrtPriceX96));
